Add files via upload
Data acquisition refers to the procedure of obtaining information by downloading or           transferring files from one location to another. File transfer, FASTA manipulation, manipulating Excel data sheets, and data management are the four primary sections of data acquisition.

Data wrangling is a time-consuming and iterative process of preparing and enriching data for analysis and visualization. Data preparation is the term for a procedure of data analysis like this. If the data wrangling system' output contains new data or errors, this procedure, like data analysis itself, can be iterative This implies that the program sequences may be repeated until the preferred outcome is obtained. Data wrangling is a simple and straightforward process compared to data curation or data stewardship, which are much more complex and difficult. Data curation is a holistic process that defines the continuous management of data throughout its whole life cycle from creation and first storage to the spot in time when it is archived or outdated and removed for future analysis.